+++ apps/web/src/app/(site)/(blog)/blog/[slug]/page.tsx
