+++ apps/web/src/app/(site)/(blog)/blog/author/[slug]/opengraph-image.tsx
