+++ apps/web/src/app/(site)/(blog)/layout.tsx
