+++ apps/web/src/app/(site)/(legal)/layout.tsx
