+++ apps/web/src/app/(site)/agent/auth/register/route.ts
