+++ apps/web/src/app/(site)/agent/auth/revoke/route.ts
