+++ apps/web/src/app/(site)/contributors/opengraph-image.tsx
