--- apps/web/src/lib/navigation/site-chrome.ts
@@ -1,5 +1,0 @@
-const LANDING_PATHNAME = "/";
-
-export function isFullBleedPath(pathname: string): boolean {
-  return pathname === LANDING_PATHNAME;
-}
