--- apps/web/src/types/site-shell.ts
@@ -1,5 +1,0 @@
-import type { ReactNode } from "react";
-
-export interface SiteShellProps {
-  children: ReactNode;
-}
